StyletIoC: Move from storing Types to storing TypeHandles
Types are reasonably expansive, and when you've got a lot of them, the
memory usage adds up. We don't actually need the Type most of the time -
the hot path just does a dictionary lookup, and we only need the
RuntimeTypeHandle for that.

Besides, Type.GetTypeFromHandle is cheap - typeof(T) does it internally.
So, just use RuntimeTypeHandle on the hot path, and swallow the cost
of an extra Type.GetTypeFromHandle in the rare case that we need an
actual Type.