Use powershell, and powershell only, for Stylet.Start
NuGet's 'content' approach broke horribly when we tried to upgrade:
it removed App.xaml.cs.

Using powershell through and through is more powerful: we can have a
package which immediately uninstalls itself, which means it doesn't
remain part of the user's package set. We can also be smarter about
upgrading the project: we don't nuke anything that's in App.xaml, we
don't need the user to manually delete MainWindow.xaml, and we can keep
anything that's currently in MainWindow.xaml