Ensure that exceptions from ActiveItem are rethrown
Conductor<T>.Collections.OneActive has its ActiveItem set from a binding
in most cases. However, bindings swallow exceptions encountered when
setting the source. This means that the user switched to a new
ViewModel, and its OnActivate/OnInitialActivate method threw an
exception (or the OnDeactivate method of the ViewModel being switched
away from threw an exception), this exception would be swallowed by the
binding.

Work around this by introducing RethrowingBinding, which explicitly
rethrows any exceptions encountered when setting the source on the
dispatcher. Use this for the ActiveItem binding on
StyletConductorTabControl.

Fixes #133